Well i made follow project OBJECTSOCIALWebsite host will be object.social That project will be same as OBJECTSOCIALSoftware.
OBJECTSOCIAL will be business project.

ClaimsSoftware are social project only for software.
Claims will be private project.
Claims have.:
MemoryClaimsWebsite this will be the main host for Claims the host will be memory.claims
In this we have BadClaimsWebsite have a host name bad.claims
In this we have GodClaimsWebsite have a host name good.claims