Now we need to have knowledge about the device. If Software we support Android, Microsoft, Mac, iOS
As we upgrade more languages, it only makes sense to know the language.

After this i have alot issue whit MAUI so i made new project for MAUI and need to test it on Mac and iOS to see if it get same error on Android. this is the icon error.